Move YAML parsing to a PluginDescription subclass.
Plugin interfaces may have other, more intuitive means of getting
metadata for plugins in the target language.

Because Commands are integral part of Bukkit, there's no need to
separate the regular YAML parsing and command parsing in classes.
A separate method will do.
